+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# 0.2.2
+# 0.3.0
++ Check if public IP is different from DNS record
++ Add browse zones test
++ Fix code smells
++ Add browse dns records test
+
+## 0.2.2
 + Fix yaml
 
 ## 0.2.1